chore: refactor api package
