+++ pkg/api/v1/handlers/task.go
@@ -2,7 +2,7 @@
 package handlers
 
 import (
-	"github.com/celestiaorg/talis/internal/api/v1/services"
+	"github.com/celestiaorg/talis/internal/services"
 	"github.com/celestiaorg/talis/internal/types/infrastructure"
 
 	fiber "github.com/gofiber/fiber/v2"
